+++ src/stores/profile.js
@@ -6062,7 +6062,6 @@ export const useProfileStore = defineStore('profile', {
 
 
               }else{
-
                 // loop till we find the first one
                 for (let pt in this.activeProfile.rt[component.parentId].pt){
                   if (this.activeProfile.rt[component.parentId].pt[pt].id == component.id){
@@ -6076,7 +6075,6 @@ export const useProfileStore = defineStore('profile', {
 
               if (ptObjFound != false){
                 console.log("Found orignal here:",ptObjFound)
-
                 if (ptObjFound.hashCode == component.hashCode){
 
                   // if the component we found in the system already has data in it then we are going to add a new component
@@ -6087,7 +6085,6 @@ export const useProfileStore = defineStore('profile', {
 
                   // we also need to create new @guid valuse inside the userValue in case they add multiple versions of the same
                   component = this.componentLibraryUpdateUserValueGuid(component)
-
 
                   if (Object.keys(ptObjFound.userValue).length <= 1 || ptObjFound.id.includes("id_loc_gov_ontologies_bibframe_adminmetadata")){
                     // if this is 1 or 0 then the userdata is empty, with just a @root property
@@ -6129,12 +6126,10 @@ export const useProfileStore = defineStore('profile', {
                       }
                     }
                     let newId = component.id + "_" + (total_components+1)
-
                     let oldId = JSON.parse(JSON.stringify(component.id))
 
                     // rename it
                     component.id = newId
-
                     // add it to the pt
                     this.activeProfile.rt[component.parentId].pt[newId] = JSON.parse(JSON.stringify(component))
                     // add it to the order
@@ -6212,6 +6207,26 @@ export const useProfileStore = defineStore('profile', {
         for (let group of this.componentLibrary.profiles[key].groups){
           if (group.id == id){
             group.label = newLabel
+            this.saveComponentLibrary()
+            return true
+          }
+        }
+      }
+    },
+
+    /**
+    * Set or unset the default nature of the component
+    */
+    makeComponentDefault(id){
+      for (let key in this.componentLibrary.profiles){
+        for (let group of this.componentLibrary.profiles[key].groups){
+          if (group.id == id){
+            if (Object.keys(group).includes("useDefault")){
+              group.useDefault = !group.useDefault
+            } else {
+              group["useDefault"] = true
+            }
+
             this.saveComponentLibrary()
             return true
           }