+++ src/views/Edit.vue
@@ -142,7 +142,7 @@
       // gives access to this.counterStore and this.userStore
       ...mapStores(usePreferenceStore,useProfileStore),
       ...mapState(usePreferenceStore, ['styleDefault','panelDisplay', 'createLayoutMode']),
-      ...mapState(useProfileStore, ['profilesLoaded','activeProfileSaved']),
+      ...mapState(useProfileStore, ['profilesLoaded','activeProfileSaved', 'returnComponentLibrary']),
 
 
       ...mapWritableState(usePreferenceStore, ['showDebugModal', 'createLayoutMode','panelSizePresets']),
@@ -215,6 +215,22 @@
           }
         }
       })
+
+
+      // if there are default component, switch'em on. This might not be the right place
+    //   this.$nextTick(()=>{
+    //     for (let groups of this.returnComponentLibrary){
+    //       for (let group of groups.groupsOrder){
+    //         let target = groups.groups[group]
+    //         if (target[0].useDefault){
+    //           for (let component of target){
+    //             let r = this.profileStore.addFromComponentLibrary(component.id)
+    //           }
+    //         }
+    //       }
+    //     }
+    // })
+